refactor: simplify media components
- Replace ResponsiveMedia/ProjectMediaCarousel with SimpleMedia/SimpleMediaCarousel
- Remove complex responsive utilities, use modern CSS breakpoints
- Keep zoom controls, fine positioning, and YouTube embedding
- Update hero, about, projects components
