+++ package.json
@@ -35,7 +35,6 @@
         "juice": "^11.0.1",
         "lucide-react": "^0.513.0",
         "next": "15.3.3",
-        "next-video": "^2.2.0",
         "payload": "^3.40.0",
         "postcss": "^8.5.4",
         "react": "^19.0.0",
